betfair_file_parser_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
betfairlightweight==2.21.0
python-dotenv==1.0.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
aiofiles==23.2.1
